new commandline tool to generate the data for the reports
